refactor to split out common `prepareEntity` code
